% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmjoin.R
\name{build_joint}
\alias{build_joint}
\title{Build a fuzzy-join joint data.frame via LLM}
\usage{
build_joint(x, y, key1, key2, ...)
}
\arguments{
\item{x}{a data.frame to be joined on the lhs.}

\item{y}{a data.frame to be joined on the rhs.}

\item{key1}{string, name of the key column of data.frame x waiting for
pairing. Must name an existing column in x.}

\item{key2}{string, name of the key column of data.frame y waiting for
pairing. Must name an existing column in y.}

\item{...}{extra params passed to chat_llm()}
}
\value{
a 2-column data.frame mapping values from key1 to key2.
}
\description{
Build a fuzzy-join joint data.frame via LLM
}
\examples{
\donttest{
  build_joint(
    x = data.frame(x = c("01","02","04")),
    y = data.frame(y = c("January","Feb","May")),
    key1 = "x", key2 = "y"
  )
}
}
