% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmjoin.R
\name{check_joint}
\alias{check_joint}
\title{ask LLM to check if the built joint is correct.}
\usage{
check_joint(.joint, ...)
}
\arguments{
\item{.joint}{2-column data.frame, the built joint.}

\item{...}{extra params passed to `chat_llm()`}
}
\description{
ask LLM to check if the built joint is correct.
}
