% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmjoin.R
\name{build_joint}
\alias{build_joint}
\title{build fuzzy-join joint}
\usage{
build_joint(x, y, key1, key2, ...)
}
\arguments{
\item{x}{a `data.frame` to be join on the lhs.}

\item{y}{a `data.frame` to be join on the rhs.}

\item{key1}{string, name of the key column of data.frame `x` waiting for paring.}

\item{key2}{string, name of the key column of data.frame `y` waiting for paring.}

\item{...}{extra params passed to `chat_llm()`}
}
\value{
the Fuzzy-joined `data.frame`
}
\description{
build fuzzy-join joint
}
\examples{
\dontrun{
  build_joint(
    x = data.frame(x = c("01","02","04")),
    y = data.frame(y = c("January","Feb","May")),
    key1 = "x", key2 = "y"
  )
}
}
