% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{chat_llm}
\alias{chat_llm}
\title{Send message to LLM server}
\usage{
chat_llm(
  .message,
  .model = NULL,
  .temperature = 0,
  .max_tokens = 30000,
  .timeout = 300,
  .verbose = getOption("llmjoin.verbose", FALSE)
)
}
\arguments{
\item{.message}{the message to send.}

\item{.model}{character, LLM model to use. By default NULL (uses config value).}

\item{.temperature}{OpenAI style randomness control (0~1), by default 0.}

\item{.max_tokens}{Max tokens to spend.}

\item{.timeout}{Max seconds to communicate with LLM.}

\item{.verbose}{logical, print progress messages. Default \code{getOption("llmjoin.verbose", FALSE)}.}
}
\value{
LLM answer, strings
}
\description{
This function send message to LLM model and retrive the result.
}
\examples{
\dontrun{
chat_llm("tell a joke.")
}
}
