% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utils.R
\name{chat_llm}
\alias{chat_llm}
\title{Send message to LLM server}
\usage{
chat_llm(.message, .model = "gpt-4.1-mini", .temperature = 0.1)
}
\arguments{
\item{.message}{the message to send.}

\item{.model}{character, LLM model to use. By default gpt-4.1-nano.}

\item{.temperature}{GPT style randomness control (0~1), by default 0.1, the larger, the more rigorous.}
}
\value{
LLM answer, strings
}
\description{
This function send message to LLM model and retrive the result.
}
\examples{
chat_llm("tell a joke.")
}
