% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_joint}
\alias{check_joint}
\title{ask LLM to check if the built joint is correct.}
\usage{
check_joint(.joint, model = NULL)
}
\arguments{
\item{.joint}{2-column data.frame, the built joint.}

\item{model}{the LLM model to use}
}
\description{
ask LLM to check if the built joint is correct.
}
