% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{joint_prompt}
\alias{joint_prompt}
\title{Generate connector prompt}
\usage{
joint_prompt(x, y)
}
\arguments{
\item{x}{1-column `data.frame` or vector of characters, left hand side of the join}

\item{y}{1-column `data.frame` or vector of characters, right hand side of the join}
}
\value{
a connectors `data.frame` for joining.
}
\description{
Generate a prompt to guide the LLM in generating a joint for dataframe joinning, leveraging the two key columns from the tables to be connected.
By far(2025/04/10), DeepSeek R1 and GPT-4o showed the best result, other LLMs might fabricate non-existing data in the result.
}
\examples{
# run in Rstudio script panel:
\dontrun{
  joint_prompt(
    data.frame(x = c("01","02","04")),
    data.frame(y = c("January","Feb","May"))
  )
}
}
