% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{llm_join}
\alias{llm_join}
\title{Fuzzy join with LLM}
\usage{
llm_join(x, y, key1, key2)
}
\arguments{
\item{x}{a `data.frame` to be join on the lhs.}

\item{y}{a `data.frame` to be join on the rhs.}

\item{key1}{string, name of the key column of data.frame `x` waiting for paring.}

\item{key2}{string, name of the key column of data.frame `y` waiting for paring.}
}
\value{
the Fuzzy-joined `data.frame`
}
\description{
Fuzzy join with LLM
}
