% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-utils.R
\name{set_llm}
\alias{set_llm}
\title{set up your LLM service}
\usage{
set_llm(url = NULL, key = NULL)
}
\arguments{
\item{url}{url to your LLM provider}

\item{key}{api-key of your service}

\item{model}{the model to use, by default `NULL`}
}
\description{
Set up your LLM service (of course you can use your local ollama).
All information will be stored in your system env and won't upload to anywhere else.
Rest assured, your privacy is secure.
}
