% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tbl2md}
\alias{tbl2md}
\title{Turning R dataframe into markdown table}
\usage{
tbl2md(tbl, nm = NULL)
}
\arguments{
\item{tbl}{a `data.frame` object or a vector.}

\item{nm}{character, only used if `tbl` is a vector.}
}
\value{
markdown style table string lines
}
\description{
Turning R dataframe into markdown table
}
\examples{
tbl2md(iris)
}
